Port to bevy

+++ src/map/rect.rs
@@ -1,12 +1,12 @@
 pub struct Rect {
-    x: usize,
-    y: usize,
-    w: usize,
-    h: usize
+    x: u32,
+    y: u32,
+    w: u32,
+    h: u32
 }
 
 impl Rect {
-    pub fn new(x: usize, y: usize, w: usize, h: usize) -> Self {
+    pub fn new(x: u32, y: u32, w: u32, h: u32) -> Self {
         Rect {
             x,
             y,
@@ -15,23 +15,23 @@ impl Rect {
         }
     }
 
-    pub fn top(&self) -> usize {
-        self.y + self.h - 1
+    pub fn top(&self) -> u32 {
+        self.y + self.h as u32 - 1
     }
 
-    pub fn left(&self) -> usize {
+    pub fn left(&self) -> u32 {
         self.x
     }
 
-    pub fn bottom(&self) -> usize {
+    pub fn bottom(&self) -> u32 {
         self.y
     }
 
-    pub fn right(&self) -> usize {
+    pub fn right(&self) -> u32 {
         self.x + self.w - 1
     }
 
-    pub fn center(&self) -> (usize, usize) {
+    pub fn center(&self) -> (u32, u32) {
         (self.x + self.w / 2, self.y + self.h / 2)
     }
 
@@ -39,6 +39,6 @@ impl Rect {
         self.x < other.x + other.w
             && self.x + self.w > other.x
             && self.y < other.y + other.h
-            && self.y + self.h  > other.y
+            && self.y + self.h > other.y
     }
 }
